implement task exit
Every task should be deleted. Place the exit function on to the stack.
The last ret instruction of the task will restore the last value on the
stack as eip.

Task can be added and removed. Use a queue for storing them.
Add a new task struct to store the dynamically allocated memory.

TODO: test what happens if the task deletes the call to the exit
function.